kernel: PAE

+++ kernel/paging.h
@@ -10,7 +10,7 @@ void alloc_frame(page_entry * page, int flags);
 void free_frame(page_entry *page);
 page_entry * get_page_entry(uintptr_t address, int make, page_directory * dir, int * called_alloc, int use_reserve);
 uint64_t page_get_phy_address(const page_entry *page);
-void set_frame_identity(page_entry * page, uintptr_t addr, int flags);
+void set_frame_identity(page_entry * page, uint64_t addr, int flags);
 
 page_directory * alloc_new_page_directory(int use_reserve);
 page_directory * clone_directory(const page_directory * src);
