+++ kernel/paging_template.h
@@ -7,7 +7,7 @@ void RENAME(alloc_frame)(page_entry * page, int flags);
 void RENAME(free_frame)(page_entry *page);
 page_entry * RENAME(get_page_entry)(uintptr_t address, int make, page_directory * dir, int * called_alloc, int use_reserve);
 uint64_t RENAME(page_get_phy_address)(const page_entry *page);
-void RENAME(set_frame_identity)(page_entry * page, uintptr_t addr, int flags);
+void RENAME(set_frame_identity)(page_entry * page, uint64_t addr, int flags);
 
 page_directory * RENAME(alloc_new_page_directory)(int use_reserve);
 page_directory * RENAME(clone_directory)(const page_directory * src);
